refactor(llm): 将聊天投递从 ai_callback 收回 product
